Fixed problems with filters take 1

+++ filesprocessing/filters/SmallerThanFilter.java
@@ -4,25 +4,33 @@ import java.io.File;
 
 public class SmallerThanFilter implements SizeFilters {
 
-    /**This is the minimum of the filter*/
-    double max;
+	/**
+	 * This is the minimum of the filter
+	 */
+	double max;
 
-    /**This is the conversion to kb*/
-    private static final int CONVERT = 1024;
+	/**
+	 * This is the conversion to kb
+	 */
+	private static final double CONVERT = 1024;
 
-    /**This is the constructor
-     * @param curMinimum recieves the minimum factor that it filters by
-     */
-    SmallerThanFilter (double curMinimum){
-        max = curMinimum/CONVERT;
-    }
+	/**
+	 * This is the constructor
+	 *
+	 * @param curMinimum recieves the minimum factor that it filters by
+	 */
+	SmallerThanFilter(double curMinimum) {
+		max = curMinimum / CONVERT;
+	}
 
-    /**This is the method that actually filters
-     * @param pathname the files path
-     * @return true if the file passes the path
-     */
-    @Override
-    public boolean accept(File pathname) {
-        return max < pathname.length();
-    }
+	/**
+	 * This is the method that actually filters
+	 *
+	 * @param pathname the files path
+	 * @return true if the file passes the path
+	 */
+	@Override
+	public boolean accept(File pathname) {
+		return max < pathname.length();
+	}
 }
